Simulate a single generation.

Defined in detail in Gaynor et al. 2023. To summarize, this function is a single time step in a stochastic stage-structured matrix population dynamics model for diploid, triploid, and autotetraploid perennial plants with two life-stages (reproductively immature and reproductively mature). Population composition at time t + 1 is defined by reproduction, survival, and maturation. Note, lists of three should always have values representing c(diploids, triploid, autotetraploids).

Usage

one.iter.f.choosy(
  ct.vec,
  aii.vec,
  as.matur,
  as.msurv.e.set,
  d,
  gnum.base,
  s,
  b,
  cc,
  mc,
  density.type = "all",
  mate.lazy = FALSE
)

Arguments

ct.vec

Population composition at time t. Indicates the sum of each type at time t, ct.vec = c(2x_immature, 3x_immature, 4x_immature, 2x_mature, 3x_mature, 4x_mature)

aii.vec

The survival probability of an immature individual for each cytotype. Must be a list of three integers between 0 and 1. For example, aii.vec = c(0.5, 0.3, 0.5).

as.matur

The probability of maturation from an immature stage to the mature stage for each cytotype. Must be a list of three integers between 0 and 1. For example, as.matur = c(0.5, 0.3, 0.5).

as.msurv.e.set

The survival probability of a mature individual for each cytotype. Must be a list of three integers between 0 and 1. For example, as.msurv.e.set = c(0.5, 0.3, 0.5). This list is defined within the gen.iter.f.choosy() function.

d

Strength of density dependency on gamete production for each cytotype. Must be a list of three integers between 0 and 1. For example, d = c(0.001, 0.009, 0.001).

gnum.base

Mean number of gametes per individual per cytotype. Must be a list of three numeric values. For example, gnum.base = c(100, 100, 100).

s

Selfing rate. Must be a single integer between 0 and 1.

b

Proportion of unreduced gamete formed by each diploid and tetraploid individual. Must be a single integer between 0 and 1.

cc

Proportion of 3n gamete formation from each triploid individual. Must be a single integer between 0 and 1.

mc

Strength of mating choice. Must be a single integer between 0 and 1.

density.type

Default = "all", this sets the density at time t as all individuals at time t. Alternatively, "like-cytotype" sets the density at time t for each cytotype based on only the total immature and mature individuals of that cytotype at time t.

mate.lazy

Default = FALSE, this prevents selfing from occurring during outcrossing. However, this increases the computational time by 31x!

Value

List of 9, with 1:6 representing the number of individuals of each cytotype and at both stages at time t + 1. Items 7:9 are the number of gametes sampled for 2x, 3x, and 4x individuals at time t.