Ives et al. 2003 stability metrics calculator.

Based on Ives et al. 2003, we calculate four stability metrics based on the estimates of the matrix of ecological interactions B and the estimated variance-covariance matrix of environmental noise 'Sigma'. These two quantities can be read from the output of the function mars.cls().

Usage

stability(B, sigma)

Arguments

B

Matrix of p x p (p = number of species), where \(b_{ij}\) gives the effect of the abundance of species j on per capita growth rate of species. This can be calculated with the mars.cls() function.

sigma

Matrix of p x p, environmental noise variance-covariance matrix. This can be calculated with the mars.cls() function.

Value

Resulting list includes:

var.prop, at stationarity the var.prop is the variance proportion attributable to environmental noise. The smaller the values, the more stable the dynamics.

mean.return.time & var.return.time, rate at which the transition distribution converges back to the stationary distribution. The less time it takes to return to the stationary distribution, the more stable the population.

reactivity, measures reaction to perturbations or the distance away from stationary a system moves in response to a disturbance. Again, smaller is better in terms of stability.

sp.contribs, squared eigenvalue representing the characteristic return rate of the variance of the transition distribution of the estimated MARS(1) Markov Process.

References

Ives, A. R., Dennis, B., Cottingham, K. L., Carpenter, S. R. (2003). Estimating community stability and ecological interactions frm time-series data. Ecological Monographs, 72(2): 301 - 330.