Simulate a single generation.

Defined in detail in Gaynor et al. 2025. To summarize, this function is a single time step in a stochastic stage-structured matrix population dynamics model for diploid, triploid, and autotetraploid perennial plants with two life-stages (reproductively immature and reproductively mature). Population composition at time t + 1 is defined by reproduction, survival, and maturation. Note, lists of three should always have values representing c(diploids, triploid, autotetraploids).

Usage

one.iter.f.choosy(
  ct.vec,
  aii.vec,
  as.matur,
  as.msurv.e.set,
  d,
  gnum.base,
  s,
  b,
  cc,
  mc,
  gam.density.type,
  is.density.type,
  mate.lazy = FALSE,
  rj = c(1, 1),
  yj = c(1, 1),
  env.sigma,
  Xi.is,
  env.immature.survival
)

Arguments

ct.vec

Population composition at time t. Indicates the sum of each type at time t, ct.vec = c(2x_immature, 3x_immature, 4x_immature, 2x_mature, 3x_mature, 4x_mature)

aii.vec

The survival probability of an immature individual for each cytotype. Must be a list of three integers between 0 and 1. For example, aii.vec = c(0.0005, 0.005, 0.0005).

as.matur

The probability of maturation from an immature stage to the mature stage for each cytotype. Must be a list of three integers between 0 and 1. For example, as.matur = c(0.60, 0.06, 0.60).

as.msurv.e.set

The survival probability of a mature individual for each cytotype. Must be a list of three integers between 0 and 1. These are defined based on user supplied mean probability of mature survival (as.msurv), proportion of environmental variance (env.ci), This list is defined within the gen.iter.f.choosy() function.

d

Strength of density dependency on gamete production for each cytotype. Must be a list of three integers between 0 and 1. For example, d = c(0.001, 0.009, 0.001).

gnum.base

Mean number of gametes per individual per cytotype. Must be a list of three numeric values. For example, gnum.base = c(100, 100, 100).

s

Selfing rate. Must be a single integer between 0 and 1.

b

Proportion of unreduced gamete formed by each diploid and tetraploid individual. Must be a single integer between 0 and 1.

cc

Proportion of 3n gamete formation from each triploid individual. Must be a single integer between 0 and 1.

mc

Strength of mating choice. Must be a single integer between 0 and 1.

gam.density.type

Default = "all", this sets the density dependence type for number of gametes produced at time t as all individuals at time t. Alternatively, "like-cytotype" sets the density at time t for each cytotype based on only the total immature and mature individuals of that cytotype at time t.

is.density.type

Default = "all", this sets the density dependence type for immature surivival at time t as all individuals at time t. Alternatively, "like-cytotype" sets the density at time t for each cytotype based on only the total immature and mature individuals of that cytotype at time t.

mate.lazy

Default = FALSE, this prevents selfing from occurring during outcrossing. However, this increases the computational time by 31x!

rj

List of two indicates the stage dependent density dependent impact of immature and mature individuals on number of gametes produced. Default is c(1,1).

yj

List of two indicates the stage dependent density dependent impact of immature and mature individuals on the probability of immature survival. Default is c(1,1).

env.sigma

Sigma value to define environmental variance corresponding to immature survival rate per generation.

Xi.is

Sample from a multivariate normal distribution. Allows immature and mature survival probabilities be correlated, as well as allowing for a correlation among cytotypes based on a supplied rho value.

env.immature.survival

Default = FALSE. When FALSE, the mean immature survival rate is equal to the exponential of the inverse immature survival rate (aii) times the sum of immature individuals and mature individuals of the cytotype indicated by is.density.type, which may be modified by yj. When this variable equals TRUE, the previous mean, or the immature determinate survival rate, is transformed by log(mu/(1-mu)) and the Johnson SB distribution is sampled given the env.sigma and Xi.is (defined in gen.iter.f.choosy()).

Value

List of 15, with 1:6 representing the number of individuals of each cytotype and at both stages at time t + 1. Items 7:9 are the number of offspring generated for 2x, 3x, and 4x individuals at time t. Items 10:12 are the number of gametes generated from 2x, 3x, and 4x individuals at time t. Items 12:15 are the immature survival probabilities at time t.